feat(ivy-extra): Enhance default action handling
Extend `ivy-extra-extra-default-actions` to support actions that
operate on symbols, with detailed documentation on how each action
is structured and how to use the alist.

Introduce `ivy-extra-extra-default-non-intern-actions` for actions
that operate directly on strings, providing an example with
`display-buffer` and its associated lambda function.

Update `ivy-extra-add-extra-actions` to handle both types of actions
and clarify the distinction between actions that expect a symbol and
those that expect a string.